shadows

+++ examples/shadows.nim
@@ -22,7 +22,7 @@ proc generateShadowImage(radius: int, offset: Vec2,
                          fillStyle: ColorRGBA = rgba(255, 255, 255, 255),
                          shadowColor: ColorRGBA = rgba(0, 0, 0, 255)
                          ): Image =
-  let sz = int(2*radius)
+  let sz = 2*radius
   let radius = radius.toFloat
 
   let circle = newImage(sz, sz)
